apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
python-pptx==1.0.2
pymupdf==1.25.3
pypdf==5.1.0
pillow==11.1.0
networkx==3.4.2
httpx==0.28.1
google-generativeai==0.8.4
google-cloud-texttospeech==2.24.0
google-cloud-storage==2.19.0
aiofiles==24.1.0
jinja2==3.1.5
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.20.0
